move components to new folder structure
