+++ packages/components/src/Combobox/components/Action/Action.css
@@ -1,0 +1,12 @@
+.action {
+  width: 100%;
+  padding: var(--space-small) var(--space-base);
+  border: none;
+  text-align: left;
+  background-color: var(--color-surface);
+}
+
+.action:hover,
+.action:focus {
+  background-color: var(--color-surface--hover);
+}
