+++ packages/components/src/Combobox/components/Action/Action.css.d.ts
@@ -1,0 +1,5 @@
+declare const styles: {
+  readonly "action": string;
+};
+export = styles;
+
