+++ packages/components/src/Combobox/components/Action/index.ts
@@ -1,0 +1,1 @@
+export { Action, ActionProps } from "./Action";
