+++ packages/components/src/Combobox/components/Content/Content.css.d.ts
@@ -1,5 +1,4 @@
 declare const styles: {
-  readonly "combobox": string;
   readonly "content": string;
   readonly "hidden": string;
   readonly "search": string;
@@ -7,10 +6,7 @@ declare const styles: {
   readonly "triggerIcon": string;
   readonly "optionsList": string;
   readonly "selectedOption": string;
-  readonly "trigger": string;
-  readonly "selection": string;
   readonly "actions": string;
-  readonly "action": string;
 };
 export = styles;
