+++ packages/components/src/Combobox/components/Content/index.ts
@@ -1,0 +1,1 @@
+export { Content } from "./Content";
