+++ packages/components/src/Combobox/components/Trigger/TriggerChip.css
@@ -1,0 +1,17 @@
+.trigger {
+  display: inline-flex;
+  gap: var(--space-smaller);
+  padding: var(--space-small);
+  border: var(--border-base) dashed var(--color-interactive);
+  border-radius: calc(var(--radius-large) * 5);
+}
+
+.triggerIcon {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: calc(var(--space-smaller) * 5);
+  height: calc(var(--space-smaller) * 5);
+  border-radius: var(--radius-circle);
+  background-color: var(--color-surface--background);
+}
