+++ packages/components/src/Combobox/components/Trigger/TriggerChip.css.d.ts
@@ -1,0 +1,6 @@
+declare const styles: {
+  readonly "trigger": string;
+  readonly "triggerIcon": string;
+};
+export = styles;
+
