+++ packages/components/src/Combobox/components/Trigger/types.ts
@@ -1,0 +1,6 @@
+export interface TriggerProps {
+  /**
+   * Trigger label
+   */
+  readonly label: string;
+}
